requirements.txt
python-chess
gpiozero
smbus2
setproctitle
rpi_ws281x
keyboard
flask